+++ src/FacebookConnector.php
@@ -8,39 +8,26 @@
  * @copyright Copyright (c) 2017 Simon Müller itsCoding
  */
 
-namespace boscho87fbconn\facebookconnector;
+namespace itscoding\facebookconnector;
 
-use boscho87fbconn\facebookconnector\services\EntryPoster as EntryPosterService;
-use boscho87fbconn\facebookconnector\services\EventFetcher as EventFetcherService;
-use boscho87fbconn\facebookconnector\services\TokenLoader as TokenLoaderService;
-use boscho87fbconn\facebookconnector\services\EntryPoster;
-use boscho87fbconn\facebookconnector\services\EventFetcher;
-use boscho87fbconn\facebookconnector\models\Settings;
-use boscho87fbconn\facebookconnector\services\TokenLoader;
-
-use boscho87fbconn\facebookconnector\widgets\OAuth;
+use itscoding\facebookconnector\services\EntryPoster as EntryPosterService;
+use itscoding\facebookconnector\services\EventFetcher as EventFetcherService;
+use itscoding\facebookconnector\services\TokenLoader as TokenLoaderService;
+use itscoding\facebookconnector\services\EntryPoster;
+use itscoding\facebookconnector\services\EventFetcher;
+use itscoding\facebookconnector\models\Settings;
+use itscoding\facebookconnector\services\TokenLoader;
+use itscoding\facebookconnector\widgets\OAuth;
 use Craft;
 use craft\base\Plugin;
 use craft\elements\Entry;
 use Craft\events\ModelEvent;
-use craft\services\Plugins;
-use craft\events\PluginEvent;
 use craft\console\Application as ConsoleApplication;
-use craft\web\UrlManager;
 use craft\services\Dashboard;
 use craft\events\RegisterComponentTypesEvent;
-use craft\events\RegisterUrlRulesEvent;
 use yii\base\Event;
 
 /**
- * Craft plugins are very much like little applications in and of themselves. We’ve made
- * it as simple as we can, but the training wheels are off. A little prior knowledge is
- * going to be required to write a plugin.
- *
- * For the purposes of the plugin docs, we’re going to assume that you know PHP and SQL,
- * as well as some semi-advanced concepts like object-oriented programming and PHP namespaces.
- *
- * https://craftcms.com/docs/plugins/introduction
  *
  * @author    Simon Müller itsCoding
  * @package   FacebookConnector
@@ -54,30 +41,16 @@ use yii\base\Event;
  */
 class FacebookConnector extends Plugin
 {
-    // Static Properties
-    // =========================================================================
 
     /**
      * Static property that is an instance of this plugin class so that it can be accessed via
      * FacebookConnector::$plugin
-     *
      * @var FacebookConnector
      */
     public static $plugin;
 
-    // Public Methods
-    // =========================================================================
-
     /**
-     * Set our $plugin static property to this class so that it can be accessed via
-     * FacebookConnector::$plugin
-     *
-     * Called after the plugin class is instantiated; do any one-time initialization
-     * here such as hooks and events.
-     *
-     * If you have a '/vendor/autoload.php' file, it will be loaded for you automatically;
-     * you do not need to load it in your init() method.
-     *
+     * executed on every request
      */
     public function init()
     {
@@ -86,15 +59,13 @@ class FacebookConnector extends Plugin
         self::$plugin = $this;
         // Add in our console commands
         if (Craft::$app instanceof ConsoleApplication) {
-            $this->controllerNamespace = 'boscho87fbconn\facebookconnector\console\controllers';
+            $this->controllerNamespace = 'itscoding\facebookconnector\console\controllers';
         }
-
         $this->setComponents([
-                'tokenLoader' => TokenLoader::class,
-                'entryPoster' => EntryPoster::class,
-                'eventFetcher' => EventFetcher::class
-            ]
-        );
+            'tokenLoader' => TokenLoader::class,
+            'entryPoster' => EntryPoster::class,
+            'eventFetcher' => EventFetcher::class
+        ]);
 
         Event::on(
             Entry::class,
@@ -106,57 +77,28 @@ class FacebookConnector extends Plugin
             }
         );
 
-        // Register our site routes
-        Event::on(
-            UrlManager::class,
-            UrlManager::EVENT_REGISTER_SITE_URL_RULES,
-            function (RegisterUrlRulesEvent $event) {
-                $event->rules['siteActionTrigger1'] = 'facebook-connector/default';
-            }
-        );
-
-        // Register our CP routes
-        Event::on(
-            UrlManager::class,
-            UrlManager::EVENT_REGISTER_CP_URL_RULES,
-            function (RegisterUrlRulesEvent $event) {
-                $event->rules['cpActionTrigger1'] = 'facebook-connector/default/do-something';
-            }
-        );
-
-
-        // Register our widgets
         Event::on(
             Dashboard::class,
             Dashboard::EVENT_REGISTER_WIDGET_TYPES,
             function (RegisterComponentTypesEvent $event) {
                 if (Craft::$app->request->get('code')) {
+                    //handle facebook callback
                     FacebookConnector::$plugin->tokenLoader->handleCallback();
                     $errors = FacebookConnector::$plugin->tokenLoader->getErrorMessages();
-                    if (count($errors) > 0) {
-                        Craft::$app->session->setError(implode(' ', $errors));
-                    } else {
-                        Craft::$app->session->setNotice('Loaded a Valid Token');
+                    if (!count($errors) > 0) {
+                        Craft::$app->session->setNotice(
+                            Craft::t('facebook-connector','Loaded a Valid Token')
+                        );
                     }
+                    Craft::$app->session->setError(implode(' ', $errors));
                 }
                 $event->types[] = OAuth::class;
             }
         );
 
-        // Do something after we're installed
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function (PluginEvent $event) {
-                if ($event->plugin === $this) {
-                    // We were just installed
-                }
-            }
-        );
-
         /**
+         *Todo remove this comment and code if its not needed anymore
          * Logging in Craft involves using one of the following methods:
-         *
          * Craft::trace(): record a message to trace how a piece of code runs. This is mainly for development use.
          * Craft::info(): record a message that conveys some useful information.
          * Craft::warning(): record a warning message that indicates something unexpected has happened.
@@ -182,12 +124,8 @@ class FacebookConnector extends Plugin
         );
     }
 
-    // Protected Methods
-    // =========================================================================
-
     /**
      * Creates and returns the model used to store the plugin’s settings.
-     *
      * @return \craft\base\Model|null
      */
     protected function createSettingsModel()
@@ -198,7 +136,6 @@ class FacebookConnector extends Plugin
     /**
      * Returns the rendered settings HTML, which will be inserted into the content
      * block on the settings page.
-     *
      * @return string The rendered settings HTML
      */
     protected function settingsHtml(): string
@@ -210,14 +147,4 @@ class FacebookConnector extends Plugin
             ]
         );
     }
-
-    /**
-     * @return string
-     */
-    public static function getBaseUrl()
-    {
-        return (isset($_SERVER['HTTPS']) ? 'https://' : 'http://') . $_SERVER['HTTP_HOST'] . '/';
-    }
-
-
 }