+++ src/models/Settings.php
@@ -8,20 +8,12 @@
  * @copyright Copyright (c) 2017 Simon Müller itsCoding
  */
 
-namespace boscho87fbconn\facebookconnector\models;
-
+namespace itscoding\facebookconnector\models;
 
 use craft\base\Model;
 
 /**
  * FacebookConnector Settings Model
- *
- * This is a model used to define the plugin's settings.
- *
- * Models are containers for data. Just about every time information is passed
- * between services, controllers, and templates in Craft, it’s passed via a model.
- *
- * https://craftcms.com/docs/plugins/models
  *
  * @author    Simon Müller itsCoding
  * @package   FacebookConnector
@@ -29,34 +21,30 @@ use craft\base\Model;
  */
 class Settings extends Model
 {
-    // Public Properties
-    // =========================================================================
+
 
     /**
-     * Some field model attribute
-     *
-     * @var string
+     * @var string application id of facebook app
      */
-
     public $appId = 'appId';
 
+    /**
+     * @var string application secret of facebook app
+     */
     public $appSecret = 'appSecret';
 
+    /**
+     * @var string pageId to post to
+     */
     public $pageId = 'pageId';
 
+    /**
+     * @var string facebook api version
+     */
     public $apiVersion = 'v2.10';
-
-    // Public Methods
-    // =========================================================================
 
     /**
      * Returns the validation rules for attributes.
-     *
-     * Validation rules are used by [[validate()]] to check if attribute values are valid.
-     * Child classes may override this method to declare different validation rules.
-     *
-     * More info: http://www.yiiframework.com/doc-2.0/guide-input-validation.html
-     *
      * @return array
      */
     public function rules()