+++ src/services/EventFetcher.php
@@ -8,21 +8,14 @@
  * @copyright Copyright (c) 2017 Simon Müller itsCoding
  */
 
-namespace boscho87fbconn\facebookconnector\services;
+namespace itscoding\facebookconnector\services;
 
-use boscho87fbconn\facebookconnector\FacebookConnector;
+use itscoding\facebookconnector\FacebookConnector;
 
 use craft\base\Component;
 
-
 /**
  * EventFetcher Service
- *
- * All of your plugin’s business logic should go in services, including saving data,
- * retrieving data, etc. They provide APIs that your controllers, template variables,
- * and other plugins can interact with.
- *
- * https://craftcms.com/docs/plugins/services
  *
  * @author    Simon Müller itsCoding
  * @package   FacebookConnector
@@ -30,28 +23,6 @@ use craft\base\Component;
  */
 class EventFetcher extends Component
 {
-    // Public Methods
-    // =========================================================================
-
-    /**
-     * This function can literally be anything you want, and you can have as many service
-     * functions as you want
-     *
-     * From any other plugin file, call it like this:
-     *
-     *     FacebookConnector::$plugin->eventFetcher->exampleService()
-     *
-     * @return mixed
-     */
-    public function exampleService()
-    {
-        $result = 'something';
-        // Check our Plugin's settings for `someAttribute`
-        if (FacebookConnector::$plugin->getSettings()->someAttribute) {
-        }
-
-        return $result;
-    }
 
 
 }