+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## v0.15.0 - 2017-10-08
+### Changed
+  - Readme Update
+
+## v0.14.0 - 2017-09-25
+### Added 
+  - First german Translations
+  - Added some lines of Doc
+
+## v0.13.0 - 2017-09-25
+### Fixed
+  - fixed codacy issues
+
+## v0.12.0 - 2017-09-25
+### Changed
+  - renamed vendor of the project
+
 ## v0.11.0 - 2017-09-25
 ### Changed
   - Moved the token callback function to the dashboard page, so no more templates are needed